"capsule fit bone" button
